requirements.txt
PyQt5
python-pptx
cx_Freeze